figs_chap6/tti-figure/pyproject.toml
[project]
name = "tti-figure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jax>=0.6.2",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "tyro>=0.9.25",
]
